consumer/requirements.txt
kafka-python
redis
langchain
langchain-ollama
pydantic<3,>=1